+++ app/src/main/java/com/example/jastip/ui/screen/user/SoreScreen.kt
@@ -72,7 +72,7 @@ fun SoreScreen(navController: NavController,
                         .align(Alignment.CenterStart)
                         .padding(start = 16.dp) // Pindah padding ke sini
                         .size(24.dp)
-                        .clickable { navController.popBackStack() }
+                        .clickable { navController.navigate("user") }
                 )
                 Text(
                     text = "SORE",
