+++ app/src/main/java/com/example/jastip/ui/screen/user/loginscreen/LoginScreen.kt
@@ -147,7 +147,7 @@ fun LoginScreen(
                 LaunchedEffect(Unit) {
                     when (viewModel.userRole) {
                         "admin" -> navController.navigate("admin")
-                        "user" -> navController.navigate("main")
+                        "user" -> navController.navigate("user")
                     }
                 }
             }
